sw
#include <omp.h>
#include <stdio.h>
#include <stdlib.h>

#define NRA 62     /* number of rows in matrix A */
#define NCA 15     /* number of columns in matrix A */
#define NCB 7      /* number of columns in matrix B */

int main (int argc, char *argv[]) 
{
   int    tid, nthreads, i, j, k, chunk;
   double    a[NRA][NCA],           /* matrix A to be multiplied */
   b[NCA][NCB],           /* matrix B to be multiplied */
   c[NRA][NCB];           /* result matrix C */

   chunk = 10;                    /* set loop iteration chunk size */

/*** Spawn a parallel region explicitly scoping all variables ***/
#pragma omp parallel shared(a,b,c,nthreads,chunk) private(tid,i,j,k)
{
  tid = omp_get_thread_num();
  if (tid == 0)
  {
    nthreads = omp_get_num_threads();
    printf("Starting matrix multiple example with %d threads\n",nthreads);
    printf("Initializing matrices...\n");
  }
  /*** Initialize matrices ***/
  #pragma omp for schedule (static, chunk) 
  
  for (i=0; i < NRA; i++)
    for (j=0; j < NCA; j++)
       a[i][j] = i + j;

#pragma omp for schedule (static, chunk)
    for (i=0; i<NCA; i++)
        for (j=0; j < NCB; j++)
            c[i][j] = 0;

  /*** Do matrix multiply sharing iterations on outer loop ***/
  /*** Display who does which iterations for demonstration purposes ***/

  printf("Thread %d starting matrix multiply...\n",tid);
  #pragma omp for schedule (static, chunk)
     for (i=0; i < NRA; i++) {
      for (j=0; j < NCB; j++)
         for (k = 0; k < NCA; k++)
           c[i][j] += a[i][k] * b[k][j];
     }
  } // end of parallel region
//**Print Results**//

  printf("******************************************************\n");
  printf("Result Matrix:\n");
  for (i=0; i < NRA; i++)
  {
    for (j = 0; j < NCB; j++)
     printf("%6.2f ", c[i][j]);
    printf("\n");
  }
  printf("******************************************************\n");
  printf("Done\n");
}

This next example involves combined parallel loop reduction. It demonstrates a sum reduction within a combined parallel loop construct. Notice that default data element scoping is assumed. There are no clauses specifying shared or private variables. OpenMP will automatically make loop index variables private within team:

#include <omp.h>
#include <stdio.h>
#include <stdlib.h>

int main (int argc, char *argv[]) 
{
    int   i, n;
    float a[100], b[100], sum; 

    /* Some initializations */
    n = 100;
    for (i=0; i < n; i++)
      a[i] = b[i] = i * 1.0;
    sum = 0.0;

#pragma omp parallel for reduction(+:sum)
  for (i=0; i < n; i++)
    sum = sum + (a[i] * b[i]);

    printf("   Sum = %f\n",sum);
}
      a[i] = b[i] = i * 1.0;



      Sum = 328350.000000

NOTE:
      a[i] = b[i] = i * 1.0;Sum = 328350.000000
      \

      \
In Calculus, when we want to compute the area underneath a curve, we draw rectangles that approach the curvilinear line. The curve is continuous, meaning that it has a tangent line at every point. That is, at every single point that comprises the curve, there is a change, a change in direction. This is also called a differential coefficient. Each rectangle has a mid-point, and the rectangle has a height and a width. The narrower the rectangles, the more of them there are to sum. This is called nearing the area. Actually, Numerical Integration is a method of computing an approximation of the area under the curve of a function, especially when the exact integral cannot be solved. For example, the value of the constant Pi can be defined by the following integral. However, rather than solve this integral exactly, we can approximate the solution by the use of numerical integration: